Adds Microphone class to handle internal sound details. Cleans up AppDelegate code
